requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.1.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dotenv==1.2.1
django-allauth==65.4.1
django-jazzmin==3.0.1
requests>=2.31.0
cryptography>=42.0.0
sqlparse==0.5.5
tzdata==2025.3
